pyproject.toml
[tool.poetry]
name = "pelican-bibliography"
version = "0.3.0"
description = "Generated bibliography items"
authors = ["Micah Smith <micahjsmith@gmail.com>"]
license = "MIT"
readme = "README.md"
keywords = ["pelican", "plugin"]
repository = "https://github.com/micahjsmith/pelican-bibliography"
documentation = "https://docs.getpelican.com"
include = ["pelican/plugins/bibliography/data/templates/research.html"]
packages = [
    { include = "pelican" },
]

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Framework :: Pelican",
    "Framework :: Pelican :: Plugins",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.urls]
"Funding" = "https://donate.getpelican.com/"
"Issue Tracker" = "https://github.com/micahjsmith/pelican-bibliography/issues"

[tool.poetry.dependencies]
python = "^3.6"
pelican = "^4.5"
citeproc-py = "^0.5.1"
pyyaml = "^5.3"
ply = "^3.11"

[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}
flake8 = "^3.8"
flake8-black = "^0.2.0"
invoke = "^1.3"
isort = "^5.4"
livereload = "^2.6"
markdown = "^3.2.2"
pytest = "^6.0"
pytest-cov = "^2.8"
pytest-pythonpath = "^0.7.3"
pytest-sugar = "^0.9.4"
Werkzeug = "^1.0"

[tool.poetry.extras]
markdown = ["markdown"]

[tool.autopub]
project-name = "Research Page"
git-username = "botpub"
git-email = "botpub@autopub.rocks"

[tool.isort]
# Maintain compatibility with Black
combine_as_imports = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3

# Sort imports within their section independent of the import type
force_sort_within_sections = true

# Designate "pelican" as separate import section
known_pelican = "pelican"
sections = "FUTURE,STDLIB,THIRDPARTY,PELICAN,FIRSTPARTY,LOCALFOLDER"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
